CLOUDP-304933 Add operationId to commands and restructure docs generation tool
