+++ Makefile
@@ -126,7 +126,7 @@ gen-mocks: ## Generate mocks
 .PHONY: gen-docs
 gen-docs: ## Generate docs for atlascli commands
 	@echo "==> Generating docs"
-	go run -ldflags "$(LINKER_FLAGS)" ./tools/docs/main.go
+	go run -ldflags "$(LINKER_FLAGS)" ./tools/docs
 
 .PHONY: build
 build: ## Generate an atlas binary in ./bin
